pyproject.toml
[project]
name = "cosmos"
version = "1.0.0"
description = "Cosmos"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
# This pyproject.toml is for dev dependencies only
# Production dependencies are managed via manifest.json for Home Assistant

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "aioresponses>=0.7.0",
    "pytest-aiohttp>=1.0.0",
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "ruff>=0.1.0",
    "voluptuous>=0.13.0",
    "taskipy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.taskipy.tasks]
format = "ruff format ."
format-check = "ruff format --check ."
lint = "ruff check --fix ."
lint-check = "ruff check ."
test = "pytest tests/ -v"
check = { cmd = "ruff format --check . && ruff check .", help = "Run all checks" }
ci = { cmd = "ruff format --check . && ruff check . && pytest tests/ -v", help = "Run all CI checks" }
